fix #9 also update documentation to show the new comparison

+++ pygeogrids/netcdf.py
@@ -118,7 +118,7 @@ def save_lonlat(filename, arrlon, arrlat, arrcell=None,
             setattr(gpi, 'valid_range', [np.min(gpivalues), np.max(gpivalues)])
             gpidirect = 0x0b
 
-        latitude = ncfile.createVariable('lat', np.dtype('float32').char,
+        latitude = ncfile.createVariable('lat', np.dtype('float64').char,
                                          dim[0])
         latitude[:] = arrlat
         setattr(latitude, 'long_name', 'Latitude')
@@ -130,7 +130,7 @@ def save_lonlat(filename, arrlon, arrlat, arrcell=None,
             londim = dim[1]
         else:
             londim = dim[0]
-        longitude = ncfile.createVariable('lon', np.dtype('float32').char,
+        longitude = ncfile.createVariable('lon', np.dtype('float64').char,
                                           londim)
         longitude[:] = arrlon
         setattr(longitude, 'long_name', 'Longitude')
